requirements.txt
absl-py==2.3.1
attrs==25.4.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
comtypes==1.4.14
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.9
mysql-connector-python==9.6.0
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
packaging==25.0
pillow==12.1.0
protobuf==3.20.3
PyAudio==0.2.14
pycparser==2.23
pygame==2.6.1
pyparsing==3.3.1
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.99
pywin32==311
six==1.17.0
sounddevice==0.5.3
SpeechRecognition==3.14.5
typing_extensions==4.15.0
Werkzeug==3.1.5
